requirements.txt
autopep8==1.5
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
idna==2.8
pycodestyle==2.5.0
pycparser==2.20
pymongo==3.10.1
python-dotenv==0.11.0
python-telegram-bot==12.8
requests==2.22.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.8
